# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tracker API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/tasks/{task_key}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Read a task",
        "description": "Пакет преемника: всё, что нужно агенту с чистым контекстом, одним вызовом.\n\nКарточка, связи с обеих сторон со статусом задачи на другой стороне, вычисляемые\nпризнаки, последняя сводка целиком, открытые вопросы и неразобранные замечания\nцеликом, опись дела и переходы по таблице. Тела остальных записей читаются отдельно\nв `GET /tasks/{key}/entries`.\nПереходы перечислены по таблице; валидации (заполненные разделы, сводка, вердикты,\nблокеры, дети) проверяются в момент перехода, а не при чтении.",
        "operationId": "read_task",
        "security": [
          {
            "ApiToken": []
          }
        ],
        "parameters": [
          {
            "name": "task_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Task key `QUEUE-number`; matching ignores case",
              "examples": [
                "TRK-42"
              ],
              "title": "Task Key"
            },
            "description": "Task key `QUEUE-number`; matching ignores case"
          },
          {
            "name": "X-Actor-Label",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Signature of a temporary agent, latin snake_case. Required with a shared agent token (one issued without a participant), ignored with a participant token",
              "examples": [
                "nightly_agent"
              ],
              "title": "X-Actor-Label"
            },
            "description": "Signature of a temporary agent, latin snake_case. Required with a shared agent token (one issued without a participant), ignored with a participant token"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_TaskPackageRead_"
                }
              }
            }
          },
          "401": {
            "description": "Token is missing, unknown or revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Action is not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Object not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "tasks"
        ],
        "summary": "Update a task",
        "description": "Меняет только переданные поля.\n\nНазвание, описание и пять разделов — только в `backlog` (иначе `409\ntask_field_locked`); исполнитель, теги и приоритет — в любом незакрытом статусе; в\n`done` и `cancelled` не меняется ничего (`409 task_closed`). Каждое изменение\nподшивает запись: раздел — `section_changed`, исполнитель — `assignee_changed`,\nтеги и приоритет — `field_changed`. Поля без записи не бывает: изменение, не\nоставившее записи, не доходит до ленты (`CONCEPT.md`, 4.1). `version` — не поле\nзадачи, а условие: устаревшая версия отвечает `409 version_conflict`.",
        "operationId": "update_task",
        "security": [
          {
            "ApiToken": []
          }
        ],
        "parameters": [
          {
            "name": "task_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Task key `QUEUE-number`; matching ignores case",
              "examples": [
                "TRK-42"
              ],
              "title": "Task Key"
            },
            "description": "Task key `QUEUE-number`; matching ignores case"
          },
          {
            "name": "X-Actor-Label",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Signature of a temporary agent, latin snake_case. Required with a shared agent token (one issued without a participant), ignored with a participant token",
              "examples": [
                "nightly_agent"
              ],
              "title": "X-Actor-Label"
            },
            "description": "Signature of a temporary agent, latin snake_case. Required with a shared agent token (one issued without a participant), ignored with a participant token"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_TaskRead_"
                }
              }
            }
          },
          "401": {
            "description": "Token is missing, unknown or revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Action is not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Object not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_key}/transition": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Move a task to another status",
        "description": "Переводит задачу по таблице переходов.\n\nПереход не из таблицы — `409 transition_not_allowed` с допустимыми целями в\n`details.allowed`. Шаг назад по цепочке и отмена требуют `reason` (`422\ntransition_reason_required`); `backlog → open` требует заполненных разделов (`422\ntask_sections_incomplete`). Выход из `in_progress` требует сводки, подшитой после\nпоследнего входа в него (`409 summary_required`); `in_progress → done` —\nположительного последнего вердикта по каждой проверке, подшитого после последнего\nвхода в `in_progress` (`409 checks_not_passed`, незасчитанные проверки в\n`details.checks` парами `check_no` и `reason`). Вход в\n`in_progress` отклоняется при открытом блокере (`409 task_blocked`, их ключи в\n`details.blockers`), переход в `done` — при детях не в `done` и не в `cancelled`\n(`409 task_has_unclosed_children`, ключи в `details.children`). Переход подшивает\n`status_changed` с `from`, `to` и `reason`.",
        "operationId": "transition_task",
        "security": [
          {
            "ApiToken": []
          }
        ],
        "parameters": [
          {
            "name": "task_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Task key `QUEUE-number`; matching ignores case",
              "examples": [
                "TRK-42"
              ],
              "title": "Task Key"
            },
            "description": "Task key `QUEUE-number`; matching ignores case"
          },
          {
            "name": "X-Actor-Label",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Signature of a temporary agent, latin snake_case. Required with a shared agent token (one issued without a participant), ignored with a participant token",
              "examples": [
                "nightly_agent"
              ],
              "title": "X-Actor-Label"
            },
            "description": "Signature of a temporary agent, latin snake_case. Required with a shared agent token (one issued without a participant), ignored with a participant token"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskTransition"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_TaskRead_"
                }
              }
            }
          },
          "401": {
            "description": "Token is missing, unknown or revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Action is not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Object not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tasks/{task_key}/entries/{entry_no}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Read one case entry",
        "description": "Одна запись по номеру внутри задачи — адрес из ссылки `TRK-42#12`.\n\nНомера, которого в задаче нет, — `404 entry_not_found`: номера выдаются без дыр,\nпоэтому промах означает ссылку на чужое дело, а не пропущенную страницу.",
        "operationId": "read_task_entry",
        "security": [
          {
            "ApiToken": []
          }
        ],
        "parameters": [
          {
            "name": "task_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Task key `QUEUE-number`; matching ignores case",
              "examples": [
                "TRK-42"
              ],
              "title": "Task Key"
            },
            "description": "Task key `QUEUE-number`; matching ignores case"
          },
          {
            "name": "entry_no",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Entry number inside the task, from 1",
              "examples": [
                12
              ],
              "title": "Entry No"
            },
            "description": "Entry number inside the task, from 1"
          },
          {
            "name": "X-Actor-Label",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Signature of a temporary agent, latin snake_case. Required with a shared agent token (one issued without a participant), ignored with a participant token",
              "examples": [
                "nightly_agent"
              ],
              "title": "X-Actor-Label"
            },
            "description": "Signature of a temporary agent, latin snake_case. Required with a shared agent token (one issued without a participant), ignored with a participant token"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_EntryRead_"
                }
              }
            }
          },
          "401": {
            "description": "Token is missing, unknown or revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Action is not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Object not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DataResponse_EntryRead_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/EntryRead"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "DataResponse[EntryRead]"
      },
...
      "DataResponse_TaskPackageRead_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TaskPackageRead"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "DataResponse[TaskPackageRead]"
      },
      "DataResponse_TaskRead_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TaskRead"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "DataResponse[TaskRead]"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "Единый формат ошибки: всё содержательное лежит под ключом `error`."
      },
...
      "TaskStatus": {
        "type": "string",
        "enum": [
          "backlog",
          "open",
          "in_progress",
          "waiting",
          "done",
          "cancelled"
        ],
        "title": "TaskStatus",
        "description": "Зашитый список статусов (`CONCEPT.md`, 3.3)."
      },
      "TaskTransition": {
        "properties": {
          "to": {
            "$ref": "#/components/schemas/TaskStatus",
            "description": "Target status",
            "examples": [
              "open"
            ]
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 65536
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason",
            "description": "Why the task moves. Required for any step back along the chain and for `cancelled`; optional otherwise. Recorded in the `status_changed` entry",
            "examples": [
              "Нужны уточнения по разделу context"
            ]
          },
          "version": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Version",
            "description": "Version the client last saw; omit it to skip the check",
            "examples": [
              3
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "to"
        ],
        "title": "TaskTransition",
        "description": "Перевод статуса по таблице переходов."
      },
...
    },
    "securitySchemes": {
      "ApiToken": {
        "type": "http",
        "description": "Tracker API token: `Authorization: Bearer trk_...`",
        "scheme": "bearer"
      }
    }
  }
}